Declare streams in support of the reconciliation mechanism for Dumps 2.0.
In support of the reconciliation mechanism for Dumps 2.0 (T358373), we want to declare a stream were a PySpark job will emit events when it has detected that a datalake table's copy of a revision is inconsistent with that of MediaWiki.

This event, which follows the page_content_change schema, will then be enriched by a downstream Flink job (T368787) that requires its own sink and error streams.

For now, we declare these stream as 'rc0' to start development.

Bug: T368755
Change-Id: I64bf1b8e0d2417626cd52bbc3bc3dea3aacbc2f4