logos: Sync config.yaml and logos.php
1. In I914e96e, custom width and height were manually set in
logos.php for arwiki tagline/wordmark. But this file is supposed
to be automatically generated.

2. In I914e96e, comments were added to the config.yaml without
running the script to update logos.php

3. The same thing for bjnwiki for Ie049200, where comments where
added/modified in config.yaml only

Now running the script (for new, unrelated patch) is reverting
the first change (to default values, because they're not found
in config.yaml) and is adding the second change to logos.php
(because it's found in config.yaml)

This patch updates config.yaml and regenerates logos.php with
the script to sync the two files and ensure unrelated patch
does not carry the unrelated changes.

5. arzwiki relies on arwiki wordmark logo, but arwiki decided to
change the size of the wordmark in the above changes. This change
was not propagated to arzwiki since it was not done in config.yaml
properly. Now that it's done, arzwiki must specify their own size,
else they'll receive the arwiki change. I have maintained the size
which they're currently using.

Follow-Up: I914e96ebdf05d2e815beb6ac2f4afa3d4cd931e0
Follow-Up: I3ac6fe804e824632fd9042b0ed80593cf6125e1e
Bug: T374430
Change-Id: If901bc477511b50401192baa50dbe22042a983ab