Turn on Parsoid Selective Update metrics
In order to evaluate the feasibility and impact of Parsoid selective
update, we collect detailed statistics on parses.  The calculation is a
little bit expensive (we do an otherwise-unnecessary cache query and
also a parse and DOM diff of the 'previous' parse of the page) so
randomly sample 1-in-1000 pages to avoid impacting user-visible
performance too much.

Bug: T371713
Depends-On: I42bbd370c4eba46de40261511cf49d7c462f5bfe
Change-Id: I8c2d6d75bc68a03a5ebc39d58f0987962f9665f6