Enable {{USERLANGUAGE}} on Commons and Meta
Enable {{USERLANGUAGE}} on Commons as an alternative to the extensive
use of {{int:lang}} there. {{int:lang}} is a hack which requires a page
of the form [[MediaWiki:Lang/LLL]] to be created for every defined
language LLL, containing the language code itself.

Also enable {{USERLANGUAGE}} on Meta, which has thousands of usages of
{{int:lang}}, including in the Community Wishlist Survey which was the
motivating use case for implementing the feature.

Also fix unclosed vim fold.

Bug: T4085
Change-Id: I2c3ae1f0ad0be76c558abe6fe7611b8a28c4ba71