Profiler: centralize metrics send to a function
Conslidates the duty of emitting metrics to a single function
emitMetrics() which handles the socket.

Update docs stats-host to include that it can be null.
While the code previously allowed to be unset/optional,
this was not documented, and is no longer the case now.
It is now a nullable but required option.

Change-Id: I96d179dc903b0f41c1fbb592fb4dc7fbfd860b9f