votewiki, testwiki: add securepoll-edit-poll to electionadmin
In Ibdb052c742d18e7422ee5c2f755fbc617d7d9ea0, actions allowed by the
extension by virtue of electionadmin group membership are changed to be
based on presence of securepoll-edit-poll right instead.

Add the right to the group here - ensuring the patches are together a
no-op on votewiki and testwiki.

Bug: T377531
Change-Id: I036fba2225dbbe2188d52cdb1fa78484483ff902