TimedMediaHandler: use shellbox globally
Now that we have isolated and fixed the issue with long-running
encodes, restore all videos to be encoded using shellbox-video. The
source of the initial problem was that shellbox does not set a tcp
keepalive and so connections were timing out at the lvs level when
they ran over 20 or so minutes.

This issue has been addressed at the service mesh level, so restore
use of the mesh to connect to shellbox rather than connecting
directly.

Bug: T357309
Change-Id: Ie40f9dd09db1905a2ecff4773b1cb97724f23374