Remove 'mainpage' from $wgForceUIMsgAsContentMsg for Wikidata
Including 'mainpage' in $wgForceUIMsgAsContentMsg allows a wiki to have
completely separate main pages per language, with the title of each
language’s main page determined by the mainpage message in that
language. However, this has not worked well on Wikidata; if the main
page is not created in that language, it will simply be a redlink in the
sidebar, which is not helpful. Instead, Wikidata has largely migrated to
a single, translatable main page at Wikidata:Main Page [1], and set most
MediaWiki:Mainpage/* pages to 'Wikidata:Main Page' to ensure that these
languages use this translatable main page. This config change makes the
latter part unnecessary, and ensures that all languages use the English
version of the mainpage message, pointing at the central translatable
main page.

For languages that already had a MediaWiki:Mainpage/* override pointing
to Wikidata:Main Page, this should make no difference (though the
override becomes redundant and can be deleted). For languages that had
no such override and did not have a separate main page, this provides
the significant benefit of actually giving them a working main page link
(Wikidata:Main Page may not be translated yet, but it is better than a
redlink, and can also be translated by users). However, there is also a
third set of languages: those where the mainpage message points to a
separate main page which does exist on Wikidata. An overview of these
two dozen or so languages is available at the current revision of
User:Nikki/Sandbox#Main page [2]. For the most part, those separate main
pages are very outdated (you can tell if they don’t include the
“network” image, which was added to the central main page in 2014). In
these cases, the separate translated main page is currently used, while
the translation of the central main page (if it exists – some do but not
all) is unreachable and unused; with this change, the translation of the
central main page is used instead, and the separate main page becomes
unreachable and its translated strings unused. This is considered an
acceptable price to pay for fixing the main page everywhere else.

[1]: https://www.wikidata.org/wiki/Wikidata:Main_Page
[2]: https://www.wikidata.org/w/index.php?title=User:Nikki/Sandbox&oldid=2269020968#Main_page

Bug: T184386
Change-Id: I0c2f3acc1d521d4ab96a428e0bc067ba2e2d65d5