DB config for testcommonswiki deployment for Charts
For test deployment of Charts with a local JsonConfig repo at testcommonswiki

* points virtual domain mapping for `globaljsonlinks`
* enables $wgTrackGlobalJsonLinks for testwiki+testwikicommons

Requires that the `globaljsonlinks`, `globaljsonlinks_target`, and
`globaljsonlinks_wiki` tables be installed from the JsonConfig extension
into testcommonswiki, expected to be a temporary deployment to be removed
when moving over to the production config.

Turning off $wgTrackGlobalJsonLinks will eliminate code paths going through
the database table, so if there's any trouble with the database mapping simply
remove all this at the possible cost of some test data missing some link
records.

Bug: T379199
Change-Id: I2e54b123ce2877b3ccddbfcd678a8055cd24c74b