Disable various extensions when using the shared login domain
As a security hardening measure, disable a bunch of
non-authentication-related extensions when using ths shared
authentication domain. Based on the existing rules for loginwiki
(via the 'lockeddown' wiki tag).

This is an alternative to the approach proposed in
Ieb24afcfc9c6326002649d60a14a40be2ca30e7e, and only one of those
should be merged.

Bug: T373737
Change-Id: If370a41e51564e2817a00da9b3f87a2a6026d856