Separate cache key space for test & production JsonConfig data
The shared cache key space for JsonConfig doesn't include a database
name so it's possible to cross-pollute between the test set and the
production deployment, in theory.

Going ahead and setting a distinct cache key for the testcommonswiki
and commonswiki JsonConfig repo deployments to prevent surprises.
Should be a harmless switchover, no prep needed.

Bug: T380320
Change-Id: I46e52949c3ed6bc26aa76b9abda7bd3e0fb7a9af