Enable canShellboxGetTempUrl everywhere
Indicate to SwiftFileBackend that it should send files to Shellbox as
tempauth URLs, rather than downloading the files locally and POSTing
them. This will improve performance by avoiding duplicate network
transmission, and avoiding the need for an HMAC hash (Shellbox requests
are cryptographically signed).

The Shellbox 4.1.0 server needs to be deployed before this can be
merged.

Bug: T292322
Change-Id: I87711b100a18ddd946def6878a70a95ba675950e