Use a bespoke database configuration for dumps
When on dumps, use a bespoke database configuration that is resolved
from SRV records. This is acceptable because the dumps are not
performance sensitive and so the dns queries (one per section, plus one)
per host in the list, are not a problem for a script running for hours
or days.

Bug: T382947
Change-Id: I2a7e78f8438cc524e97ff4b92a6e5f3871d00587