Disable distinct-values constraint checks on Commons
These have never worked, as an unintentional side effect of the way
WikibaseQualityConstraints writes its SPARQL queries. We’re about to
change those SPARQL queries in a way which might result in the query
service starting to return some results on Commons; however, those
results would not be useful: they would only list other entities with
the same value on Wikidata, but (absent a configured Commons Query
Service) not on Commons. We might eventually fix this later (I have no
idea), but for now let’s ensure we don’t see any unintentional results,
by completely disabling the constraint type on Commons, just like we
previously disabled type constraints: it’s a hack, but it works for now.

Also reformat the comments to be nicer.

Bug: T369079
Change-Id: Ibdbe941e3d129e6b96ee20e547fbbdcf513b2452