Remove `tmpAlwaysShowMulLanguageCode` temporary setting
The `tmpAlwaysShowMulLanguageCode` / `wbTmpAlwaysShowMulLanguageCode`
setting was using during the soft rollout of the `mul` / default
language feature to allow us to test the behaviour of the
fully-enabled feature on test wikis before we rolled out the full
feature.

Now that `mul` is released, we can remove this setting and
associated behaviours.

Bug: T330217
Depends-On: I1a68e3916be1f96e149d2c549930802ad4a1e0c7
Change-Id: I27cd944c5f017b3785af19c5ff67ce53fe16a869