Do not disable extensions on SUL3 shared authentication domain
It turned out to be extremely disruptive, as various caching
strategies (e.g. the extension registry or community configuration)
assume that the set of enabled extensions is constant for a given
wiki. LocalUserCreated hooks not getting called was also problematic.

Disabling extensions seems unsustainable even for a very small set of
extensions (see T373737#10508131 for details). Since it was only
defense in depth, and arguably it added more security risks than it
removed, we'll just have to do without it.

Bug: T373737
Bug: T384919
Bug: T384236
Change-Id: I5c05875a189d6656b7757dc09a5135c3ef88d184