Remove /pt from ptwikibooks $wgUploadMissingFileUrl
This dates back to the initial public commit (015f5b7131), so it’s hard
to know for sure why it’s there, but I assume it was just a simple
mistake and isn’t needed anymore. Let’s remove it for consistency. (As
far as I can tell, it does absolutely nothing.)

Change-Id: I8675dce6aff0e638f622d287b173512a876fe2c7