Add sourceswiki to $wgImportSources for all Wikisources
Add the multilingual Wikisource (sourceswiki) as a default import
source for all Wikisources. The multilingual Wikisource functions
as the Incubator for Wikisources, so it is only logical that it
is a default import source for them all.

Also add it to all Wikisources that had specific lists of import
sources, and standardize those to use the 'mul' interwiki prefix
instead of 'oldwikisource'.

Bug: T385591
Change-Id: I5ed9435655a031e672c7374badbbb056d7b998db