tests: Move buildLogoHTML.php to tests/ alongside buildConfigCache.php
This file is unrelated to Multiversion. It is neither in charge of
routing, nor does it consume Multiversion routing (e.g. MWScript.php).

This patch is in support of single version containers (T384944), to
make the code clearer in terms of what is part of multiversion and
thus what needs migration or replacement long-term.

The only call to MWWikiversions in this local test script, is a call
to `array_keys(wikiversions.json)` to obtain a list of all wkis.
That's fine but does not warrant the file to be "part" of multiversion.
This is actually identical to MWWikiversions::readDbListFile('all'),
as enforced by MWWikiversionsTest::testWikiversionsFileComplete.

Reading of dblists also does not depend on or bear any relation to
multiversion, which I'll move out in a separate patch.

Change-Id: I0cf91231109d77c147e5c1fde0933db111237c2c