Remove temporary '-php8' and '-k8s' suffixes from ArcLamp pipeline
This change removes the conditional logic that appended -k8s or -php8 to
the excimer Redis channel (T288165). With the migrations to Kubernetes
and PHP8 both complete, there is no longer a need to distinguish these
samples. A follow-up change to operations/puppet will decommission the
k8s- and php8-specific ArcLamp consumers.

Bug: T391516
Change-Id: I138a591224f2899236bf1349a758d99bf9e5e343