Enable fixed Wikibase RDF on Beta
This makes the hashes of blank nodes for somevalue (unknown value) snaks
use the URI that the v: prefix resolves to, rather than the prefix name
itself (i.e. "v"), and then renames the prefix from v: to wdv: to match
the query service. Only enable it in Beta for now; Test Wikidata and
then real Wikidata will follow later, coordinated with an announcement.

Bug: T384344
Change-Id: I1fccbde426e117cb959c44570cca9716786dec0b
Depends-On: I5e9da89e5096146898f9ac087a75942990868d5e