[Growth] Set default api lookahead size to 10
Increases the size of pages used for fetching
tasks from the growthtasks api. This is to improve
the prediction of downstream results left to return
by the API. The discrepancy between the API total
count and the actual results is introduced (mainly) by
the protected pages filter ProtectionFilter in
GrowthExperiments as a way to handle the limitation
from CirrusSearch, see T259346.

Bug: T325990
Change-Id: Iea257a46197c5665868463be67cc80e06b7a6578