Introduce config setting to disable default event-organizer group
Add a Wikimedia-specific config flag to configure whether a wiki should
use the default event-organizer group created in T376822. If not, unset
the group in a 'MediaWikiServices' hook handler, after the
`GroupPermission` value has been merged from extension.json (and not in
an extension function as that won't likely work due to T275334).

For beta, use the new setting instead of the previous code that no
longer works since the event-organizer group was defined in
extension.json.

Bug: T386290
Change-Id: Id52b08b074c6bbb98d8917cd95e74802350c1e75