beta: Do not undeclare wmgGEActiveExperiment
SiteConfiguration ignores any config set to `null`, meaning setting
a config to `null` in InitialiseSettings.php or InitialiseSettings-labs.php
is actually an instruction to UNSET the variable, rather than an instruction
to set it to null.

CommonSettings.php assumes $wmgGEActiveExperiment exists
in all cases (the correct value for "no experiment" is false,
which is set by InitialiseSettings.php, rather than null).
To meet that assumption, we need to stop undeclaring the variable
for no-experiment beta wikis.

Bug: T386846
Change-Id: Ibe59228feac6e0f098134a0898c811c05b379681