When executing cli scripts, wait for the service mesh
In some conditions, a script might be executed while the service mesh is
unavailable or unhealthy, which results in undefined behaviour for most
scripts that need to enqueue any type of event or job, or that need to
reach other services.

Therefore, add a test for the mesh health before we execute the script.
The possibiliyt of bypassing the check is also provided for use in
emergencies and/or in loops like foreachwiki.

Bug: T387208
Change-Id: I67724eedd21d8456b762c8c94b304e6bef90584b