Deduplicate JsonConfig config
It's so duplicated, even the word 'config' is duplicated.

$wgJsonConfigs is now written to in only three places, so one can
figure out what is the final value without mental acrobatics.
Values that are identical for every config entry, but different on
every wiki, are shared in $wmgJsonConfigDataModeConfig.
The config is mostly declared in InitialiseSettings instead of
having many special cases based on $wgDBname.

The generated configuration should be exactly the same as before.

Also move doc comments to a better place, and add more docs.

Change-Id: I32458b547b6acc103f0b66c91e07a6eb026ffbc2