docroot: Enable Chrome credential sharing on all open SUL wikis
Generated with a temporary "make-auth-assetlinks-json.php" script
I wrote (https://phabricator.wikimedia.org/P73922), executed locally
from the operations/mediawiki-config directory.

Note that this is a one-time file that, while it needs to be kept
indefinitely, it does not need to be updated in the future to include
new wikis.

This file exists to instruct the password manager in Google Chrome,
to autofill credentials saved in the past from any of these domains,
when logging in via the auth.wikimedia.org domain.

Future wikis will be created in the SUL3 era, and thus Chrome users
will be both saving and retreiving their passwords from the same
domain (auth.wikimedia.org), which does not involve the local domain
and thus doesn't require sharing to or from the auth domain.

Bug: T385520
Change-Id: Ibe3900c9729b5be6e127a7aed9c0522645a1bd6d