Drop $wmgCampaignEventsProgramsAndEventsDashboardEnabled
The current implementation of tracking tools does not allow disabling a
tool for "writes" only. Therefore, if a tool is disabled, we will crash
when trying to display an event that uses that tool (which might have
been added on another wiki).

Ideally, we would always use the same list of tools everywhere, and only
allow disabling whether a given tool can be *added* to events on the
current wiki. However, this decoupling would require at least an
additional config settings, and possibly more generalizations to do it
well (see details in T387025#10608132).

Since this is just an edge case in our current configuration, simply
re-enable the Programs & Events Dashboard integration everywhere.

Bug: T387025
Change-Id: If9f550213f1a06ff64bb027452b1aeb39f10c56d