logos: have CI fail on uncommited logos.php changes
wmf-config/logos.php is generated and has the following statement near
the top:

 # This file is automatically generated. DO NOT EDIT IT BY HAND.

When patches are sent, people not necessarily regenerate the logos.php
and it tends to drift apart.

Add a new test environment to tox which generate the code, then run git
diff to verify whether any modifications might be missing.

Bug: T341412
Depends-On: I5a6e874a70bb82b1ae74f4c1a8ce65a1dec064ea
Depends-On: Ifebc9e7cb583adb329bcfd4de479e7dab64969e3
Change-Id: I2d65bfebd9a7616a09ce35f663909d33e588290a