fatal-error: Ensure action=cache max-age is higher than response time
Follows-up I1c677ca1cf7d318a9. Make this test not useless by having
the object not start in an already-expired state, which would be
the same as max-age=0.

In accordance with RFC 7234 [1], Varnish/ATS takes the response time
into account when calculating the age, so this means a response that
takes 11 seconds to arrive, and declares max-age=10, is effectively
expired on arrival. From the perspective of a proxy, there is no
distinction processing time and transit time [2], and realistically
it is unlikely that an application even could promise that a response
is fresh to the last moment it spent generating the response. More
likely than not, the backend started with the most important query,
and then spent the rest of the time querying associated data and
rendering based on that first internal query.

[1]: https://www.rfc-editor.org/rfc/rfc7234#section-4.2.3
[2]: https://en.wikipedia.org/wiki/Two_Generals'_Problem

Change-Id: I40b6bd99eadb1a8cb4b36a42426096a24f24a42a