Re-enable wgTrackGlobalJsonLinksNamespaces for JsonConf
Schema changes have been deployed and this should be safe to deploy now.
This enables use of the namespace text fields on the shared table
globaljsonlinks_wiki, and migration will be automatic as pages get edited.

This process can be allowed to run naturally or accelerated with a
forced null-edit bot.

Without this setting, namespace numbers were recorded of foreign
pages using Data: via JsonConfig and Charts, but not their namespace
texts. This could cause failures in displaying the titles in some
cases such as namespaces that are not present, or proper localization
or per-wiki customization of namespaces. Saving namespaces in the db
allows correct linking in all edge cases, but still saves space by
merging common namespace records on the same wiki.

Bug: T385917
Change-Id: I2d38eeb4535e688c559a1c6bd8522281b684d5bc