cirrus: explicitly route search traffic to eqiad
As part of migrating the search cluster in codfw to opensearch we would
like to reduce the search traffic to a bare minimum by only keeping
write requests there.
Reason is that in a mixed opensearch/elasticsearch cluster we can't
guarantee that hot shards like enwiki will have enough replicas to
handle the usual traffic.

Bug: T388610
Change-Id: Ie4e4d44e6a53741b6851b7e4a18470b1327904da