cirrus: explicitly route search traffic to codfw
As part of migrating the search cluster in eqiad to opensearch we would
like to reduce the search traffic to a bare minimum by only keeping
write requests there.
Reason is that in a mixed opensearch/elasticsearch cluster we can't
guarantee that hot shards like enwiki will have enough replicas to
handle the usual traffic.

Bug: T388610
Change-Id: I4d98ea3a4e906f3eaf2d938a0f86931220b73581