wikitech: Remove $wgCookieDomain override
For SUL2 login, it doesn't do anything (there is some difference
between setting a cookie's domain to the current domain vs.
leaving it unset, but it doesn't matter if the site does not have
subdomains). SUL3 login breaks, because it's happening on a
different domain and so the Set-Cookie header gets discarded by
the browser as invalid.

Bug: T389318
Change-Id: I43774e0c7b8053fc4ef4cc0945f17fb5c5ca209f