wdqs: enable hive/hdfs ingestion for rdf update streams
We want to store this data in hdfs to increase the retention.
Reason is that with wikidata RDF dumps being unstable we might run into
situations where the retention in kafka is no longer enough to be able
to run a data-reload on latest dump.
We don't have the tooling to use this data in hdfs but if a disaster
occurs we should find a way to use it.

Bug: T388372
Change-Id: I773517ced8ecc1ac7993568d1839dd27cae48c48