Fix badpass logging for locally nonexistent users
UserFactory will return a UserIdentity when the user does not
exist, but UserIdentityLookup doesn't. This was changed in
I0f9715256ffaa90175e06577fee3baae9c606efa, probably accidentally.
Restore it so we record failed login attempts for accounts which
don't exist don't exist on the given wiki. (This is important
since a correct login attempt would work, so these are relevant
for tracking brute-force etc.)

goodpass is not affected since by the point a success is logged,
the account exists already.

Change-Id: I5483270a94b8a91130bc6cd07225cace83502824