Use wikidata familly in $wgCirrusSearchSimilarityProfile
On I882e847229b286233977ce46c27fa2365ffa6b97, Timo pointed that the
previously used `wikidata` is a group of Wikidata wikis and thus covers
both `wikidatawiki` and `testwikidatawiki` wikis.

Revert back to use the `wikidata` group.
Remove a duplicate the duplicated `testwikidatawiki` entries.

While at it I am converting from double to single quotes to align with
the rest of the files.

Change-Id: I84d9e17609b9a9a2ed0c8ffce4f7fbca1b18bc4b