cirrus: Add services for read operations
CirrusSearch needs the ability to write to specific clusters in specific
datacenters, but reads don't need to be so specific. Add a `dnsdisc` cluster as
well as matching services for each cirrus cluster that read traffic can be
pointed at, to allow delegating the decision about what cluster to read from to
the etcd based dns much like other services.

This patch only configures the services, it does not start using
from them.

Bug: T143553
Depends-On: Ie6dfb586f6b22867a13b8b29d920da8409e94015
Change-Id: I26d1b8eb883cef7d5e3b875bca498d2ba75dbb64