Temporarily disable Lua profiler
This should reduce the rate of emergency timeout exceptions.

The linked bug relates to an exception being thrown shortly after the
deletion of a short-period POSIX timer. I think Scribunto is
responsible for most short-period timer deletions in production, so
this should be effective while I work on a fix in the relevant C code.

Bug: T389734
Change-Id: I30c3a34b579d384f801ae75fe608045a16ee234e