wikimaniawiki: add extendedconfirmed to translationadmin
This adds extendedconfirmed rights to translationadmin
as new TAs are typically onboarded as new editors on
this wiki. Since they can now edit admin protected
pages, they should be able to edit ec protected pages.
Extendedconfirmed rights is not given to TAs in the
previous commit.
This is a follow-up of
I19b7eb19d71435ac22196c893350cff35410e659.

Bug: T389729
Change-Id: I1817e6c26efaa25b6550a7042dfb60cdfe8739d1