Use `sul` dblist in InitialiseSettings
Use the newly added `sul` dblist to replace config where we were
previously using `default` to set the SUL/CentralAuth wiki state and
then overriding that for both `fishbowl` and `private` dblists.

The prior norm was functional, but required remembering that both
`fishbowl` and `private` are non-SUL/CentralAuth groups. The new norm
becomes using `default` to set the non-SUL/CentralAuth config and `sul`
to explicitly set the SUL/CentralAuth wiki config.

The `sul` dblist cannot be used in cases where the config wants
different settings for an overlapping list such as `lockeddown` or
`closed`. This is an unfortunate condition of dblist config expansion,
but happily we have a test that will catch that case when folks
accidentally trip over it.

Depends-On: If3eb85d78f8941a60bfe19fc4e9b7c4d3cc1ecc7
Change-Id: I203ebb8f433ed0af496cb0ddc9c13e438dc109cc