manage-dblist: Rename to manage-dblist.php
So that phpcs will lint it like our other PHP files. Most users are
hopefully calling it through composer anyway (hence the changes to the
help output; compare e.g. the “do not edit directly” comments in the
dblist files), so this shouldn’t be a breaking change for them.

Bug: T392819
Change-Id: I65761bce8f446c82a5aa556588d6d30c13bf9bee