multiversion: Update readDbListFile() calls from alias to WmfConfig
== Background ==

See I0cf9123110. TLDR: Historically we used to put all source
code in this repo that wasn't a config file in the multiversion
directory. The code for stuff that we use inside wmf-config at runtime
(after "multiversion" is done, and regular MediaWiki begins) has been
moved to the WmfConfig class. This makes it clear which bits of infra
code depend on Multiversion (and thus need to be thought about in
relation to the single-version idea), and which bits of code are
unrelated, needed either way, and fine as-is.

== What ==

In I0cf9123110  we moved MWMultiVersion::readDbListFile to the
WmfConfig class, with a wrapper/alias method left behind for compat,
and to minimize the diff.

This patch updates the callers. It still leaves the compat behind for
private code, Puppet, operations/dumps scripts, people's muscle memory
from eval.php, and anything else we might have missed.

Change-Id: Iedc63f41bd2a2cbee2fa9a616894e355477e8cca