Charts phase 1 deployment
Enables Chart extension to Wikipedias, with exceptions for the wikis
to be rolled out later in phases 2 and 3.

Private/fishbowl wikis are excluded as wmgEnableJsonConfigDataMode
is disabled there.

Additional non-Wikipedia wikis also to come in future rollouts.

Added dblists charts-phase1, 2, and 3 with the rollout lists for
easy reference. Final future rollout will change the default and
remove these references.

Bug: T393517
Change-Id: I180f152552391a48162c72f621318b686010b4c7