ores-extension: enable ores extention for rrla without the UI
In this patch we enable the ORES extension for idwiki adding the revertrisk language agnostic model.
We just want to enable the extension without the UI filters so that we can perform a backfill for the scores.
In a subsequent patch we will set the wgOresUiEnabled[idwiki] to true and enable the filters in the UI as well
So here we have:
 * add to dblist by running `composer manage-dblist add idwiki ores`
 * enable rrla model and disable default models
 * add the filter threshold for idwiki
 * set the wgOresUiEnabled to false

Bug: T382171
Change-Id: Ic58d56f01c8055ce4a1974e111b21f6a35b37ad9