Enable Lua transform switch for Charts on test and beta
Enable the underlying Lua transform features for JsonConfig and the
high-level Charts feature using it to make code-customizable data sets,
enabled on Labs and disabled in production as:

* wgJsonConfigTransformsEnabled - needed on store and clients
* wgChartTransformsEnabled - needed on clients

in InitialiseSettings.

Transforms currently default to off so if the patches go out ahead
of config it will be fine.

Switch these back off if necessary for an emergency disabling of the
transform feature, eg if:

* action=jsontransform API traffic is too high
* something is horribly wrong with the transform rendering
* something is unsafe or whatever

Bug: T395516
Change-Id: Ia490c1f07e66b0d098a412277b6e76805d3329ee