multiversion: Re-use prod for beta setSiteInfoForWiki
The beta branch and the prod branch did the same thing in all cases,
but was needlessly written differently.

In prod:
* Default site = 'wikipedia' (above the if/else switches)
* If not wikimedia, set site=family from domain name (e.g. wiktionary)
* If wikimedia and wikimediaSubdomains, set site=family='wikimedia' (e.g. chapter)
* Implied: If wikimedia and not in wikimediaSubdomains, keep default
  because its a special wiki like meta, which use site=wikipedia/suffix=wiki.

In beta:
* Default site = 'wikipedia' (above the if/else switches)
* If wikimedia, set site='wikipedia'
* Else, set site=family from domain name (e.g. wiktionary)

Note, that there are no chapter wikis in Beta. The only wikimedia.org
wikis in beta are special wikis. The two do the same, and it was
only simpler because it could be, not because it behaved different.

Bug: T289318
Change-Id: Ie87527e4a49286c41414c2e98b31455cf2af18f7