Empty change to test scap Depends-On handling
This change depends on a MediaWiki core change which is included in
wmf/1.45.0-wmf.7 but not wmf/1.45.0-wmf.6. As group2 is, as of this
writing, still on wmf.6, scap should warn about the missing dependency
before deploying the config change. This change serves the dual purpose
of demonstrating this feature and also checking that it still works
(there was a complaint about it in #wikimedia-operations on 2025-06-17,
see [1] ca. 13:45 and 14:35 UTC).

[1]: https://wm-bot.wmcloud.org/browser/index.php?start=06%2F17%2F2025&end=06%2F17%2F2025&display=%23wikimedia-operations

Change-Id: Iec543bee7576035521c4731bd3f3260dc2ddb384
Depends-On: I3187e82479f1d02b5f423ee18269e3f1f67d2575