Growth: enable new way of refreshing LinkRecommendations for more wikis
This enables the iterative way of refreshing LinkRecommendations for
more wikis. That maintenance script is split across database-shards and
so the set of wikis has been chosen to be distributed across shards as
well.

However, since this is touching enwiki, it should only enabled when the
work for T386867 has concluded (including running the revalidation
script).

Bug: T386250
Bug: T392944
Change-Id: I18e46aad92a60aea7644d9a726771d3477f4d83e