beta: Include allowance for wmcloud.org in wgGraphAllowedDomains
This is presumably unused given Graph disable-ment.

I'll take the liberty to brown-out/deprecate 'http', by limiting it
to wmflabs.org. If by the end of T289318, it hasn't come up that it
needs wmcloud.org, we can remove it along with most other
non-wmcloud.org refs.

We can always add it back at a later time, of course.

Bug: T289318
Change-Id: I6acc77020451a7f00b64d9ee5ebb4466ad7744b9