ParserCache: Enable purgePeriod for SqlBagOStuff
This will queue a deferred update on 1/10 requests that write to the
ParserCache, to purge 20 expired rows. This is the default for SqlBag,
and is how MainStash DB works in production today already.

This was previously disabled for 10+ years at WMF, in favor of the
`purgeParserCache.php` maintenance script that we ran daily instead.

Bug: T398806
Change-Id: I4730a9b8435473e94904f38d9b77f8287ec52733