EventStreamConfig: add the maps.tiles_change_bookworm stream
The maps backend for Tegola and Kartotherian (Postgres on bare
metal nodes) is being upgraded to Bookworm. This means a new
version of Postgres, and a brand new database bootstrapped from
scratch. We need to regenerate the Tegola's tile cache from
a brand new and clean topic/stream, to avoid impacting
the current one.

We chose to add the _bookworm suffix since every time the maps
backend will be upgraded, a new stream will need to be created
to support the new data on the database.

Bug: T381565
Change-Id: Ic42cf10c0730224dafc2bfadbe15f06e5bd8107d