Move special wikis outside of the 'wikipedia' group
This patch adds configuration overrides for all these wikis so as to
minimize the config diff and reassure folks that nothing is actually
changing. These overrides can then be cleaned up later. See comments for
more information on each.

One exception to this is wmgMonologChannels, for which I don't know how
to restore the previous value. The new value seems to make sense anyway.

Bug: T183549
Change-Id: I2d1df1285ddecfa01c3b4057815f3399200bbf41