Explicitly set wgServer etc. for private wikis under the 'wikipedia' dblist
Although for historical reasons we considered these private wikis to be
extensions of Wikipedias, this set-up is more confusing than it is
beneficial, so we're going to split them out. For now, let's explicitly
set the wgServer and wgCanonicalServer values (note how they're
different from the modern form under wikimedia.org).

Bug: T183549
Change-Id: Id29450178413cf5bea1cd77dbf06a9aae1dffae5