Add phan and use it to detect duplicated array keys
Add our standard phan config for libraries, make it check the wmf-config
directory, but for now report duplicated array keys only. This could
later be expanded to report more things, but it's quite tricky because
of 1) this repo dealing entirely with globals, and 2) MW itself not
being available for analysis.

Change-Id: I1093ea85c2912565ef36eb39ea4e71997479ee2d