build: Fix failing `phpcs` in CI on commits updating interwiki.php
To reproduce locally, check out I1cbf92392 in this repo locally
and run `composer phpcs`, compared to `JENKINS_URL=foo composer phpcs`,
the latter of which invokes bootstrap-ci.php and passes the result of
`git show HEAD --name-only` to `$config->files`.

I'm not sure why this hasn't failed CI before, but upstream
acknowledges that this doesn't work so I guess the bigger question is
not why it fails now, but why it worked before.

As a workaround, we can make it a regular pattern and accept the
hypothetical that, if a file named "dblists-index.php" or the
subdir+filename "wmf-config/interwiki.php" one day appears in a nested
subtree of this repository, that they too will be excluded.

Ref https://github.com/PHPCSStandards/PHP_CodeSniffer/issues/863

Change-Id: Ib55d810001c4608c7f6efff4471380f7a82b6439