Add CommunityConfigurationExample to extension-list
This causes the i18n to be built in both beta
and production. We do not need the messages built
in prod (there is no intention to deploy the extension
to prod), but it is not possible to build messages only in beta.

This patch can be only merged once **at least two**
deployment branches exist in CommunityConfigurationExample.

Bug: T372049
Change-Id: I4b9bf2373a663c384c6ddf0ab47435f279d8c977