Disable MobileFrontend on thankyou.wikipedia.org and nostalgia.wikipedia.org
I considered adding these as exeptions in MobileUrlCallback.php instead.

For thankyou.wikipedia.org that would still be unexpected because it
was specifically requested to not install MF there (as opposed to
using MF without redirect).

For nostalgia.wikipedia.org it would be invisible if we kept it,
because there is no redirect on this domain, and the Nostalgia skin
does not render the MF footer opt-in link. Either way Minerva
would be unsuitable for the Nostalgia content which is expected to
render with the Nostalgia skin (and is surprisingly mobile-friendly).

Bug: T400855
Bug: T152882
Change-Id: I0f6689e9a2099d5afa1f5e43feac50aae3d5cb33