WmfConfig: Document why 'preinstall' is indexed
Follows-up I47adcb776c (5ac7da0990), which added this here but it
is never used as part of wiki creation. For example, around activation
of tlwikisource (T388639, I6cd243c893) it is added and then removed
without ever being used inside InitialiseSettings.

It looked like a remnant. E.g. maybe we imagined overriding settings
for preinstall wikis, or maybe DB_LISTS was mistaken for a list
of all dblists. It is used, in MWMultiVersion::isPreInstall, which
is called on every request to decide whether to fatal, and so requires
the index.

Change-Id: I05e2205a355739504cae1010aa43511989deafe1