[Growth] wikidata: Preconfigure for limited Growth features release
The Growth team is looking into the possibility of releasing
some Growth features on Wikidata. This patch prepares
for that by preventively disabling several features
that we already know not to be compatible with Wikidata:

  * Help panel
  * Welcome survey
  * Suggested edits

Those features are disabled in the production configuration
files, even though the imminent release is only going
to happen in beta. This is done to make the future production
deployment easier – once we like how Beta is configured,
the remaining step should be only "enable production",
rather than having to copy config from beta to production first.

This patch is a no-op, as the GrowthExperiments extension
stays disabled in both beta and production.

This patch intentionally uses `wikidata` rather than `wikidatawiki`.
This is because we have two similar wikis (wikidatawiki and testwikidatawiki),
both of which will need pretty much the same configuration. Let's
use the dblist as we have it available.

Bug: T400937
Change-Id: I4bcaa40a6271fcf14751238df7e4d65337c99fc8