session: Enable MultiBackendSessionStore on `group0` wikis
Technically, this is should be a low-risk because both session
types will be `kask-session` which points to the same Cassandra
backend. This is a stepping stone patch to move anon session
type to a possibly different backend.

If everything works out fine on group0 as we monitor closely, we
will advance to `group1`, then rollout to `group2` afterwards.

Bug: T402808
Depends-On: I10101c8b928a1236eeb4d338ad16a1cd3db44bed
Change-Id: Ic9f2386a110c9ceb2d2f53eb78b6388c05c514d3