session: Enable MultiBackendSessionStore on `group1` wikis
Technically, this is should be a low-risk because both session
types will be `kask-session` which points to the same Cassandra
infrastructure. This is a stepping stone patch to move anon
session type to a possibly different backend.

We will activate this on group2 wikis if everything is fine with
group0 and group1.

Bug: T402808
Change-Id: I4444e60c09e360336e9405ab565bef991669e6a6