session: Enable MultiBackendSessionStore on `group2` wikis
Technically, this is should be a low-risk because both session
types will be `kask-session` which points to the same Cassandra
infrastructure. This is a stepping stone patch to move anon
session type to a possibly different backend.

A next natural step after this if it runs with no issues in prod
is to figure out what to point anon cache type to and use it.

Bug: T402808
Change-Id: I770ed5e8f0cbad898e3ad3144dcb4e0544d33f0d